Delete DaemonStatus.swift — GatewayConnectionManager absorbs all state
GatewayConnectionManager is now the single ObservableObject for
connection state and daemon metadata. DaemonStatus.swift is deleted.

What moved to GatewayConnectionManager:
- All @Published state (isConnected, httpPort, daemonVersion, etc.)
- SSE message pre-processing (handleServerMessage)
- Version compatibility checking
- Callback wiring (now in init, updates own state directly)
- AuthError enum
- ObservableObject conformance

What was removed:
- DaemonStatus class (277 lines deleted)
- All forwarding methods and accessors
- Callback pattern (onConnectionStateChanged, onDaemonVersionChanged,
  onAuthError) — GatewayConnectionManager updates its own properties

Backward compatibility preserved via typealiases:
- DaemonClient = GatewayConnectionManager
- DaemonStatus = GatewayConnectionManager
- DaemonClientProtocol kept as a protocol conformance
