+++ clients/macos/vellum-assistantTests/ChatMediaEmbedBaselineTests.swift
@@ -17,9 +17,9 @@ final class ChatMediaEmbedBaselineTests: XCTestCase {
 
     override func setUp() {
         super.setUp()
-        let cm = GatewayConnectionManager(); daemonClient = DaemonClient(connectionManager: cm)
+        daemonClient = GatewayConnectionManager()
         daemonClient.isConnected = true
-        viewModel = ChatViewModel(daemonClient: daemonClient, eventStreamClient: cm.eventStreamClient)
+        viewModel = ChatViewModel(daemonClient: daemonClient, eventStreamClient: daemonClient.eventStreamClient)
     }
 
     override func tearDown() {
