+++ clients/macos/vellum-assistantTests/ChatViewModelTests.swift
@@ -10,11 +10,11 @@ final class ChatViewModelTests: XCTestCase {
 
     override func setUp() {
         super.setUp()
-        let cm = GatewayConnectionManager(); daemonClient = DaemonClient(connectionManager: cm)
+        daemonClient = GatewayConnectionManager()
         // Mark as connected so send-path tests don't hit the disconnected guard.
         // Tests that verify disconnected behaviour explicitly set isConnected = false.
         daemonClient.isConnected = true
-        viewModel = ChatViewModel(daemonClient: daemonClient, eventStreamClient: cm.eventStreamClient)
+        viewModel = ChatViewModel(daemonClient: daemonClient, eventStreamClient: daemonClient.eventStreamClient)
     }
 
     override func tearDown() {
