+++ clients/macos/vellum-assistantTests/ConversationManagerUnseenStateTests.swift
@@ -5,7 +5,6 @@ import XCTest
 @MainActor
 final class ConversationManagerUnseenStateTests: XCTestCase {
 
-    private var cm: GatewayConnectionManager!
     private var daemonClient: DaemonClient!
     private var conversationManager: ConversationManager!
     private var sentMessages: [Any] = []
@@ -27,17 +26,15 @@ final class ConversationManagerUnseenStateTests: XCTestCase {
 
     override func setUp() {
         super.setUp()
-        cm = GatewayConnectionManager()
-        daemonClient = DaemonClient(connectionManager: cm)
+        daemonClient = DaemonClient()
         daemonClient.isConnected = true
-        conversationManager = ConversationManager(daemonClient: daemonClient, eventStreamClient: cm.eventStreamClient)
+        conversationManager = ConversationManager(daemonClient: daemonClient, eventStreamClient: daemonClient.eventStreamClient)
         conversationManager.createConversation()
     }
 
     override func tearDown() {
         conversationManager = nil
         daemonClient = nil
-        cm = nil
         sentMessages = []
         super.tearDown()
     }
