+++ clients/macos/vellum-assistantTests/PrivateConversationPersistenceTests.swift
@@ -15,11 +15,10 @@ struct PrivateConversationPersistenceTests {
     /// restorer reconstructs it as a private conversation.
     @Test @MainActor
     func privateConversationCreatedAndRestoredAsPrivate() {
-        let cm = GatewayConnectionManager()
-        let dc = DaemonClient(connectionManager: cm)
+        let dc = DaemonClient()
         dc.isConnected = true
 
-        let manager = ConversationManager(daemonClient: dc, eventStreamClient: cm.eventStreamClient)
+        let manager = ConversationManager(daemonClient: dc, eventStreamClient: dc.eventStreamClient)
 
         // ConversationManager.init enters draft mode — conversations array is empty,
         // the initial chat lives as a draftViewModel.
@@ -51,8 +50,8 @@ struct PrivateConversationPersistenceTests {
         // Now simulate a fresh restore: build a conversation list response
         // that includes this conversation with conversationType "private", and verify
         // the restorer creates it with .private kind.
-        let restorer = ConversationRestorer(daemonClient: dc, eventStreamClient: cm.eventStreamClient)
-        let delegate = MockConversationRestorerDelegate(daemonClient: dc, eventStreamClient: cm.eventStreamClient)
+        let restorer = ConversationRestorer(daemonClient: dc, eventStreamClient: dc.eventStreamClient)
+        let delegate = MockConversationRestorerDelegate(daemonClient: dc, eventStreamClient: dc.eventStreamClient)
         restorer.delegate = delegate
 
         let defaultConversation = ConversationModel()
@@ -86,17 +85,16 @@ struct PrivateConversationPersistenceTests {
     /// create and restore (control case for the private conversation test above).
     @Test @MainActor
     func standardConversationCreatedAndRestoredAsStandard() {
-        let cm = GatewayConnectionManager()
-        let dc = DaemonClient(connectionManager: cm)
+        let dc = DaemonClient()
         dc.isConnected = true
 
-        let manager = ConversationManager(daemonClient: dc, eventStreamClient: cm.eventStreamClient)
+        let manager = ConversationManager(daemonClient: dc, eventStreamClient: dc.eventStreamClient)
         // ConversationManager.init enters draft mode — no conversations in the array yet
         #expect(manager.conversations.isEmpty)
 
         // Restore a conversation list with a standard conversation
-        let restorer = ConversationRestorer(daemonClient: dc, eventStreamClient: cm.eventStreamClient)
-        let delegate = MockConversationRestorerDelegate(daemonClient: dc, eventStreamClient: cm.eventStreamClient)
+        let restorer = ConversationRestorer(daemonClient: dc, eventStreamClient: dc.eventStreamClient)
+        let delegate = MockConversationRestorerDelegate(daemonClient: dc, eventStreamClient: dc.eventStreamClient)
         restorer.delegate = delegate
 
         let defaultConversation = ConversationModel()
@@ -130,10 +128,9 @@ struct PrivateConversationPersistenceTests {
     /// restores each with the correct kind.
     @Test @MainActor
     func mixedConversationTypesRestoreCorrectly() {
-        let cm = GatewayConnectionManager()
-        let dc = DaemonClient(connectionManager: cm)
-        let restorer = ConversationRestorer(daemonClient: dc, eventStreamClient: cm.eventStreamClient)
-        let delegate = MockConversationRestorerDelegate(daemonClient: dc, eventStreamClient: cm.eventStreamClient)
+        let dc = DaemonClient()
+        let restorer = ConversationRestorer(daemonClient: dc, eventStreamClient: dc.eventStreamClient)
+        let delegate = MockConversationRestorerDelegate(daemonClient: dc, eventStreamClient: dc.eventStreamClient)
         restorer.delegate = delegate
 
         let defaultConversation = ConversationModel()
@@ -164,10 +161,9 @@ struct PrivateConversationPersistenceTests {
     /// list responses. Verify that these conversations default to .standard.
     @Test @MainActor
     func legacyPayloadWithoutConversationTypeDefaultsToStandard() {
-        let cm = GatewayConnectionManager()
-        let dc = DaemonClient(connectionManager: cm)
-        let restorer = ConversationRestorer(daemonClient: dc, eventStreamClient: cm.eventStreamClient)
-        let delegate = MockConversationRestorerDelegate(daemonClient: dc, eventStreamClient: cm.eventStreamClient)
+        let dc = DaemonClient()
+        let restorer = ConversationRestorer(daemonClient: dc, eventStreamClient: dc.eventStreamClient)
+        let delegate = MockConversationRestorerDelegate(daemonClient: dc, eventStreamClient: dc.eventStreamClient)
         restorer.delegate = delegate
 
         let defaultConversation = ConversationModel()
@@ -194,10 +190,9 @@ struct PrivateConversationPersistenceTests {
     /// (with conversationType) conversations restores correctly.
     @Test @MainActor
     func mixedLegacyAndModernPayloadsRestoreCorrectly() {
-        let cm = GatewayConnectionManager()
-        let dc = DaemonClient(connectionManager: cm)
-        let restorer = ConversationRestorer(daemonClient: dc, eventStreamClient: cm.eventStreamClient)
-        let delegate = MockConversationRestorerDelegate(daemonClient: dc, eventStreamClient: cm.eventStreamClient)
+        let dc = DaemonClient()
+        let restorer = ConversationRestorer(daemonClient: dc, eventStreamClient: dc.eventStreamClient)
+        let delegate = MockConversationRestorerDelegate(daemonClient: dc, eventStreamClient: dc.eventStreamClient)
         restorer.delegate = delegate
 
         let defaultConversation = ConversationModel()
@@ -227,11 +222,10 @@ struct PrivateConversationPersistenceTests {
     /// before the user sends any messages.
     @Test @MainActor
     func privateConversationPersistsImmediatelyViaConversationCreate() {
-        let cm = GatewayConnectionManager()
-        let dc = DaemonClient(connectionManager: cm)
+        let dc = DaemonClient()
         dc.isConnected = true
 
-        let manager = ConversationManager(daemonClient: dc, eventStreamClient: cm.eventStreamClient)
+        let manager = ConversationManager(daemonClient: dc, eventStreamClient: dc.eventStreamClient)
         manager.createPrivateConversation()
 
         let vm = manager.activeViewModel!
@@ -248,11 +242,10 @@ struct PrivateConversationPersistenceTests {
     /// callback — the kind should remain .private after the daemon responds.
     @Test @MainActor
     func privateConversationKindSurvivesIdBackfill() {
-        let cm = GatewayConnectionManager()
-        let dc = DaemonClient(connectionManager: cm)
+        let dc = DaemonClient()
         dc.isConnected = true
 
-        let manager = ConversationManager(daemonClient: dc, eventStreamClient: cm.eventStreamClient)
+        let manager = ConversationManager(daemonClient: dc, eventStreamClient: dc.eventStreamClient)
         manager.createPrivateConversation()
 
         let privateConversation = manager.conversations.first!
